+++ NEWS.md
@@ -50,6 +50,7 @@
 * Fixed suggesting filename with trailing dot when export filter is unset (by Sid, #4368)
 * Fixed snapping mode sync across instances (by Sid, #4364)
 * Fixed missing error message when 'Export as Image' fails (by kunal649, #4397)
+* Scripting: Added isExpanded/setExpanded API to LayersView and ObjectsView (#3997)
 * Scripting: Added API for custom property types (with dogboydog, #3971)
 * Scripting: Added `TileMap.chunkSize` and `TileMap.compressionLevel` properties
 * Scripting: Added optional defaultValue and toolTip params to `Dialog` add widget methods (by Oval, #4358)